Fix a bug preventing non-founders from rebooting a dead group from save
Sanctions are not saved to the tox save file, so when a peer tried to re-create
a dead group, others were not able to sync with them due to the inability
to validate the sanctions credentials. Now in such a situation we simply
don't send sanctions until a moderator or founder arrives.